requirements.txt
flask
pdfminer.six
transformers
pyyaml
psycopg2-binary
torch